requirements.txt
aniso8601==4.1.0
awsebcli==3.14.10
backports.ssl-match-hostname==3.7.0.1
blessed==1.15.0
botocore==1.12.86
cached-property==1.5.1
cement==2.8.2
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.3.9
coverage==4.5.2
decorator==4.4.0
docker==3.7.0
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
Flask==1.0.2
Flask-Cors==3.0.7
flask-restplus==0.12.1
gunicorn==20.0.4
idna==2.7
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
networkx==2.2
nose2==0.8.0
numpy==1.16.2
pathspec==0.5.5
python-dateutil==2.7.5
pytz==2018.9
PyYAML==3.13
requests==2.20.1
semantic-version==2.5.0
six==1.11.0
termcolor==1.1.0
texttable==0.9.1
typing==3.6.6
urllib3==1.24.2
wcwidth==0.1.7
websocket-client==0.54.0
Werkzeug==0.15.3
wrapt==1.11.1
